battle

+++ Unit5_WritingClasses/examples/HeroVillainExamples/Power.java
@@ -2,17 +2,17 @@ package Unit5_WritingClasses.examples.HeroVillainExamples;
 
 public class Power {
     private String power;
-    private int strength;
+    private double strength;
 
-    public Power(String power, int strength) {
+    public Power(String power, double strength) {
         this.power = power;
         this.strength = strength;
     }
 
     public String getPower() { return power; }
     public void setPower(String power) { this.power = power; }
-    public int getStrength() { return strength; }
-    public void setStrength(int strength) { this.strength = strength; }
+    public double getStrength() { return strength; }
+    public void setStrength(double strength) { this.strength = strength; }
 
     public String toString() { return "Power:" + power + "\n" + "Strength:" + strength ; }
 }
